delete post button for post-details page

+++ post-details/post-details.js
@@ -47,6 +47,9 @@ const createPost = (post) => {
             <img src="../assets/pin.svg">
             <p>${post.location}</p>
             </div>
+            <div class="post-actions">
+                <!--- EDIT/DELETE WITH PERMISSION --->
+            </div>
             <div class="image">
                 <img src="${url + '/uploads/' + post.image}">
             </div>
@@ -65,6 +68,13 @@ const createPost = (post) => {
                     <p class="description-text">${post.description}</p>
                 </div>
             </div>`
+
+    if(post.author === user.user_id) {
+        const postActionButtonsDiv = document.querySelector('.post-actions');
+        postActionButtonsDiv.innerHTML += `<button class="post-delete-btn"><a>DELETE</a></button>
+                                            <button class="post-edit-btn"><a>EDIT</a></button>`;
+    }
+
 }
 
 //Create author div
@@ -186,20 +196,23 @@ const createComments = (comments) => {
         comments.forEach((comment) =>  {
             if(user.user_id === comment.user_id || user.role_id === 0) {
                 document.getElementById(`delete${comment.comment_id}`).addEventListener('click', async (evt) => {
-                    const commentId = evt.target.id.replace(/\D/g, "");
-                    const fetchOptions = {
-                        method: 'DELETE',
-                        headers: {
-                            Authorization: "Bearer " + sessionStorage.getItem("token"),
-                        },
-                    };
-                    try {
-                        const response = await fetch(url + '/comment/' + commentId, fetchOptions);
-                        const json = await response.json();
-                        alert('Comment deleted successfully!');
-                        location.reload();
-                    } catch (e) {
-                        console.log(e.message);
+                    const answer = window.confirm("Are you sure you'd like to delete this comment?");
+                    if(answer){
+                        const commentId = evt.target.id.replace(/\D/g, "");
+                        const fetchOptions = {
+                            method: 'DELETE',
+                            headers: {
+                                Authorization: "Bearer " + sessionStorage.getItem("token"),
+                            },
+                        };
+                        try {
+                            const response = await fetch(url + '/comment/' + commentId, fetchOptions);
+                            const json = await response.json();
+                            alert('Comment deleted successfully!');
+                            location.reload();
+                        } catch (e) {
+                            console.log(e.message);
+                        }
                     }
                 })
             }
@@ -316,6 +329,33 @@ const favoriteUnfavorite = (() => {
     })
 })
 
+const editDeletePost = ((post_id) => {
+    document.querySelector('.post-edit-btn').addEventListener('click', (() => {
+        //TODO - edit post page
+        alert('GO DO EDIT PAGE');
+    }))
+
+    document.querySelector('.post-delete-btn').addEventListener('click', (async () => {
+        const answer = confirm("Are you sure you'd like to delete this post?");
+        if(answer){
+            try {
+                const fetchOptions = {
+                    method: 'DELETE',
+                    headers: {
+                        Authorization: "Bearer " + sessionStorage.getItem("token"),
+                    },
+                };
+                const response = await fetch(url + "/post/" + post_id, fetchOptions);
+                const deleted = await response.json();
+                window.location.replace(`../userpage/userpage.html?id=${user.user_id}`);
+            } catch (e) {
+                console.log(e.message);
+                alert('Error deleting post');
+            }
+        }
+    }))
+})
+
 // AJAX calls
 const getPost = async (post_id) => {
     try {
@@ -340,6 +380,7 @@ const getPost = async (post_id) => {
         addCommentForm();
         likeUnlike();
         favoriteUnfavorite();
+        editDeletePost(post_id);
     } catch (e) {
         console.log(e.message);
     }
